# src/main/resources/public/swagger.yaml
openapi: "3.0.0"
info:
  version: 1.0.0
  title: Projekt Bus
  description: Application made as university project (Java Spring Rest MongoDB)
servers:
  - url: https://peaceful-sierra-14544.herokuapp.com/
paths:
  /users:
    get:
      tags:
        - users
      summary: |
        Get list of users
      operationId: get-users
      responses:
        200:
          description: Returns Users list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/User'
    put:
      tags:
        - users
      summary: Change password
      operationId: changePassword
      requestBody:
        description: ChangePasswordTemplate object
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ChangePasswordTemplate'
      responses:
        200:
          description: Password changed
        400:
          description: Validation errors
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApplicationError'
              example: {
                "errors": [
                  "Nie znaleziono użytkownika"
                ]
              }
  /users/{login}:
    get:
      tags:
        - users
      summary: Get user by user login or email
      operationId: getUserByLogin
      parameters:
        - name: login
          in: path
          description: User login or email.
          required: true
          schema:
            type: string
      responses:
        200:
          description: User response
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        404:
          description: User not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApplicationError'
              example: {
                "errors": [
                  "Nie znaleziono użytkownika"
                ]
              }
    delete:
      tags:
        - users
      summary: Delete User by login
      operationId: deleteUserByLogin
      parameters:
        - name: login
          in: path
          description: User login
          required: true
          schema:
            type: string
      responses:
        200:
          description: User deleted
        404:
          description: User not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApplicationError'
              example: {
                "errors": [
                  "Nie znaleziono użytkownika"
                ]
              }
  /register:
    post:
      tags:
        - register
      summary: Register new User
      operationId: registerUser
      requestBody:
        description: User object
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/NewUser'
      responses:
        201:
          description: User created
          headers:
            Location:
              schema:
                type: "string"
              description: URL adress of created User
        400:
          description: Validation errors
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApplicationError'
              example: {
                "errors": [
                  "password: Haslo musi zawierac znak specjalny",
                  "password: Haslo musi zawierac liczbe",
                  "login: Dlugosc musi byc pomiedzy 5 a 32 liter"
                ]
              }
  /login:
    post:
      tags:
        - login
      summary: Login to application
      operationId: login
      requestBody:
        description: Login field may contain user login or email
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginTemplate'
      responses:
        200:
          description: User logged successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApplicationResponse'
              example: {
                "response": "eyJhbGciOiJIUzUxMiJ9.eyJzdWIiOiJhYWFsYSIsImlhdCI6MTU1NzA2NjE3OSwiZXhwIjoxNTU3MDczMzc5fQ.uCh5N4xNapNG9hLQDMTi_7BcFii0bb0Uq4Lsm7dTz8uEFMnDZQ075jK60s4y1D_yaT-gp34nno-8Nh78QpKYGQ"
              }
        404:
          description: Wrong credentionals
          content:
            application/json:
              schema:
                oneOf:
                  - $ref: '#/components/schemas/WrongPassword'
                  - $ref: '#/components/schemas/WrongLogin'
              example: {
                "response": "Zły login"
              }
  /remind-password{login}:
    post:
      tags:
        - remind-password
      summary: Remind User password
      operationId: remindPassword
      parameters:
        - name: login
          in: path
          description: User login or email.
          required: true
          schema:
            type: string
      responses:
        200:
          description: Remind successful
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApplicationResponse'
              example: {
                "response": "Mail został wysłany"
              }
        404:
          description: Wrong credentionals
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApplicationError'
              example: {
                "errors": [
                  "Nie znaleziono użytkownika"
                ]
              }
...
  /confirm-account:
    post:
      tags:
        - from-email
      summary: Confirm account from mail link
      operationId: confirmAccount
      parameters:
        - in: query
          name: tokenCode
          schema:
            type: string
          description: Confirmation token
          example: a5118be6-c8f3-4976-bc70- 8e19f776d7a5
      responses:
        200:
          description: User confirmed
        404:
          description: Token not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApplicationError'
              example: {
                "errors": [
                  "Token not found"
                ]
              }
        409:
          description: Account already confirmed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApplicationError'
              example: {
                "errors": [
                  "Account already confirmed"
                ]
              }
  /remind-password:
    post:
      tags:
        - from-email
      summary: Remind password from mail link
      operationId: remindPasswordEmail
      parameters:
        - in: query
          name: tokenCode
          schema:
            type: string
          description: Reset token code
          example: a5118be6-c8f3-4976-bc70- 8e19f776d7a5
      requestBody:
        description: PasswordTemplate object
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PasswordTemplate'
      responses:
        200:
          description: User confirmed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApplicationResponse'
              example: {
                "response": "Przystanek poprawnie usunięty z linii"
              }
        400:
          description: Validation errors
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApplicationError'
              example: {
                "errors": [
                  "password: Haslo musi zawierac litere",
                  "password: Haslo musi zawierac liczbe"
                ]
              }
        409:
          description: Password already changed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApplicationError'
              example: {
                "errors": [
                  "Hasło zostalo już zmienione"
                ]
              }
...
components:
  schemas:
    User:
      properties:
        id:
          type: string
        enabled:
          type: boolean
        login:
          type: string
          minLength: 5
          maxLength: 32
        email:
          type: string
          minLength: 5
          maxLength: 32
        password:
          type: string
          minLength: 5
          maxLength: 32
    NewUser:
      required:
        - login
        - email
        - password
      properties:
        login:
          type: string
          minLength: 5
          maxLength: 32
        email:
          type: string
          minLength: 5
          maxLength: 32
        password:
          type: string
          minLength: 5
          maxLength: 32
...
    LoginTemplate:
      required:
        - login
        - password
      properties:
        login:
          type: string
        password:
          type: string
    ChangePasswordTemplate:
      required:
        - login
        - password
        - newPassword
      properties:
        login:
          type: string
        password:
          type: string
        newPassword:
          type: string
          minLength: 5
          maxLength: 32
    PasswordTemplate:
      required:
        - password
      properties:
        password:
          type: string
          minLength: 5
          maxLength: 32
    ApplicationResponse:
      properties:
        response:
          type: string
    ApplicationError:
      properties:
        errors:
          $ref: '#/components/schemas/ErrorList'
    ErrorList:
      type: array
      items:
        properties:
          field:
            type: string
    WrongPassword:
      properties:
        message:
          type: string
    WrongLogin:
      properties:
        message:
          type: string